Adiciona políticas de permissão nas ações do controller de vacosas

+++ app/Http/Controllers/VacosasController.php
@@ -23,55 +23,66 @@ class VacosasController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create', Vacosa::class);
+
         return view('vacosas.create');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param  \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(Request $request)
     {
-        //
+        $this->authorize('create', Vacosa::class);
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  Vacosa  $vacosa
+     * @param  Vacosa $vacosa
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function show(Vacosa $vacosa)
     {
+        $this->authorize('view', Vacosa::class);
+
         return view('vacosas.show', compact('vacosa'));
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  Vacosa  $vacosa
+     * @param  Vacosa $vacosa
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(Vacosa $vacosa)
     {
+        $this->authorize('update', Vacosa::class);
+
         return view('vacosa.edit', compact('vacosa'));
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  Vacosa  $vacosa
+     * @param  \Illuminate\Http\Request $request
+     * @param  Vacosa $vacosa
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(Request $request, Vacosa $vacosa)
     {
-        //
+        $this->authorize('update', Vacosa::class);
     }
 
     /**
@@ -83,6 +94,8 @@ class VacosasController extends Controller
      */
     public function destroy(Vacosa $vacosa)
     {
+        $this->authorize('delete', Vacosa::class);
+
         return $vacosa->delete()
             ? back()->with('status', 'Vacosa excluída com sucesso!')
             : back()->with('error', 'Erro ao excluir vacosa!');
